--- a/astropy/coordinates/builtin_frames/cirs.py
+++ b/astropy/coordinates/builtin_frames/cirs.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-from astropy import units as u
 from astropy.utils.decorators import format_doc
 from astropy.coordinates.attributes import (TimeAttribute,
                                             EarthLocationAttribute)
 from astropy.coordinates.baseframe import base_doc
-from astropy.coordinates.earth import EarthLocation
 from .baseradec import doc_components, BaseRADecFrame
-from .utils import DEFAULT_OBSTIME
+from .utils import DEFAULT_OBSTIME, EARTH_CENTER
 
 __all__ = ['CIRS']
 
@@ -36,7 +34,7 @@ class CIRS(BaseRADecFrame):
     """
 
     obstime = TimeAttribute(default=DEFAULT_OBSTIME)
-    location = EarthLocationAttribute(default=EarthLocation(0*u.km, 0*u.km, 0*u.km))
+    location = EarthLocationAttribute(default=EARTH_CENTER)
 
 # The "self-transform" is defined in icrs_cirs_transformations.py, because in
 # the current implementation it goes through ICRS (like GCRS)