--- a/astropy/coordinates/builtin_frames/cirs_observed_transforms.py
+++ b/astropy/coordinates/builtin_frames/cirs_observed_transforms.py
@@ -41,7 +41,7 @@ def cirs_to_altaz(cirs_coo, altaz_frame):
     cirs_dec = usrepr.lat.to_value(u.radian)
 
     # first set up the astrometry context for CIRS<->AltAz
-    astrom = erfa_astrom.get().apio13(altaz_frame)
+    astrom = erfa_astrom.get().apio(altaz_frame)
     az, zen, _, _, _ = erfa.atioq(cirs_ra, cirs_dec, astrom)
 
     if is_unitspherical:
@@ -64,7 +64,7 @@ def altaz_to_cirs(altaz_coo, cirs_frame):
     zen = PIOVER2 - usrepr.lat.to_value(u.radian)
 
     # first set up the astrometry context for ICRS<->CIRS at the altaz_coo time
-    astrom = erfa_astrom.get().apio13(altaz_coo)
+    astrom = erfa_astrom.get().apio(altaz_coo)
 
     # the 'A' indicates zen/az inputs
     cirs_ra, cirs_dec = erfa.atoiq('A', az, zen, astrom)*u.radian