--- a/astropy/coordinates/builtin_frames/cirs_observed_transforms.py
+++ b/astropy/coordinates/builtin_frames/cirs_observed_transforms.py
@@ -22,11 +22,13 @@ from ..erfa_astrom import erfa_astrom
 
 @frame_transform_graph.transform(FunctionTransformWithFiniteDifference, CIRS, AltAz)
 def cirs_to_altaz(cirs_coo, altaz_frame):
-    if np.any(cirs_coo.obstime != altaz_frame.obstime):
-        # the only frame attribute for the current CIRS is the obstime, but this
-        # would need to be updated if a future change allowed specifying an
-        # Earth location algorithm or something
-        cirs_coo = cirs_coo.transform_to(CIRS(obstime=altaz_frame.obstime))
+
+    obsgeoloc, obsgeovel = altaz_frame.location.get_gcrs_posvel(altaz_frame.obstime)
+    if (np.any(cirs_coo.obstime != altaz_frame.obstime) or
+            np.any(cirs_coo.obsgeoloc != obsgeoloc)):
+        cirs_coo = cirs_coo.transform_to(CIRS(obstime=altaz_frame.obstime,
+                                              obsgeoloc=obsgeoloc,
+                                              obsgeovel=obsgeovel))
 
     # we use the same obstime everywhere now that we know they're the same
     obstime = cirs_coo.obstime
@@ -35,22 +37,15 @@ def cirs_to_altaz(cirs_coo, altaz_frame):
     is_unitspherical = (isinstance(cirs_coo.data, UnitSphericalRepresentation) or
                         cirs_coo.cartesian.x.unit == u.one)
 
-    if is_unitspherical:
-        usrepr = cirs_coo.represent_as(UnitSphericalRepresentation)
-        cirs_ra = usrepr.lon.to_value(u.radian)
-        cirs_dec = usrepr.lat.to_value(u.radian)
-    else:
-        # compute an "astrometric" ra/dec -i.e., the direction of the
-        # displacement vector from the observer to the target in CIRS
-        loccirs = altaz_frame.location.get_itrs(cirs_coo.obstime).transform_to(cirs_coo)
-        diffrepr = (cirs_coo.cartesian - loccirs.cartesian).represent_as(UnitSphericalRepresentation)
-
-        cirs_ra = diffrepr.lon.to_value(u.radian)
-        cirs_dec = diffrepr.lat.to_value(u.radian)
+    # We used to do "astrometric" corrections here, but these are no longer necesssary
+    # CIRS has proper topocentric behaviour
+    usrepr = cirs_coo.represent_as(UnitSphericalRepresentation)
+    cirs_ra = usrepr.lon.to_value(u.radian)
+    cirs_dec = usrepr.lat.to_value(u.radian)
 
     # first set up the astrometry context for CIRS<->AltAz
     astrom = erfa_astrom.get().apio13(altaz_frame)
-
+    astrom['diurab'] = 0
     az, zen, _, _, _ = erfa.atioq(cirs_ra, cirs_dec, astrom)
 
     if is_unitspherical:
@@ -77,23 +72,20 @@ def altaz_to_cirs(altaz_coo, cirs_frame):
 
     # first set up the astrometry context for ICRS<->CIRS at the altaz_coo time
     astrom = erfa_astrom.get().apio13(altaz_coo)
+    astrom['diurab'] = 0
 
     # the 'A' indicates zen/az inputs
     cirs_ra, cirs_dec = erfa.atoiq('A', az, zen, astrom)*u.radian
+    obsgeoloc, obsgeovel = altaz_coo.location.get_gcrs_posvel(altaz_coo.obstime)
     if isinstance(altaz_coo.data, UnitSphericalRepresentation) or altaz_coo.cartesian.x.unit == u.one:
-        cirs_at_aa_time = CIRS(ra=cirs_ra, dec=cirs_dec, distance=None,
-                               obstime=altaz_coo.obstime)
+        distance = None
     else:
-        # treat the output of atoiq as an "astrometric" RA/DEC, so to get the
-        # actual RA/Dec from the observers vantage point, we have to reverse
-        # the vector operation of cirs_to_altaz (see there for more detail)
+        distance = altaz_coo.distance
 
-        loccirs = altaz_coo.location.get_itrs(altaz_coo.obstime).transform_to(cirs_frame)
-
-        astrometric_rep = SphericalRepresentation(lon=cirs_ra, lat=cirs_dec,
-                                                  distance=altaz_coo.distance)
-        newrepr = astrometric_rep + loccirs.cartesian
-        cirs_at_aa_time = CIRS(newrepr, obstime=altaz_coo.obstime)
+    cirs_at_aa_time = CIRS(ra=cirs_ra, dec=cirs_dec, distance=distance,
+                           obstime=altaz_coo.obstime,
+                           obsgeoloc=obsgeoloc,
+                           obsgeovel=obsgeovel)
 
     # this final transform may be a no-op if the obstimes are the same
     return cirs_at_aa_time.transform_to(cirs_frame)