… CHANGES.rst
… astropy/coordinates/builtin_frames/utils.py
… astropy/coordinates/earth.py
… astropy/coordinates/erfa_astrom.py
… astropy/coordinates/tests/test_erfa_astrom.py
… astropy/coordinates/tests/test_intermediate_transformations.py
… astropy/coordinates/tests/test_regression.py
… astropy/coordinates/tests/test_sky_coord.py
… docs/coordinates/common_errors.rst
… docs/coordinates/frames.rst
… docs/coordinates/index.rst
… docs/coordinates/satellites.rst