@pep8speaks 2020-11-05
Hello @StuartLittlefair :wave:! Thanks for opening this pull request, we are very grateful for your contribution! I'm a friendly :robot: that checks for style issues in this pull request, since this project follows the [PEP8](https://www.python.org/dev/peps/pep-0008/) style guidelines. I've listed some small issues I found below, but please don't hesitate to ask if any of them are unclear!












* In the file [`astropy/coordinates/tests/test_celestial_transformations.py`](https://github.com/astropy/astropy/blob/30a953e159f2fa4c120404312379a60534f74151/astropy/coordinates/tests/test_celestial_transformations.py):

> [Line 345:1](https://github.com/astropy/astropy/blob/30a953e159f2fa4c120404312379a60534f74151/astropy/coordinates/tests/test_celestial_transformations.py#L345): [W391](https://duckduckgo.com/?q=pep8%20W391) blank line at end of file