@maxnoe 2020-11-05
I just want to comment that while I think this is great, I also think this should not make it into a release without the faster, direct ICRS to AltAz transformation.

A new release with slower transformation between ICRS and AltAz to gain some arcseconds for close-by objects would be catastrophic for us who are mostly interested in fast, not terribly precises ICRS to AltAz transforms for FAR away objects.