@mkbrewer 2020-11-05
> The cost of this precision is a rather severe slowdown in the transformation code - approximately a factor of 10. 

I am curious about this slowdown. Is this in the ICRS<->CIRS or CIRS<->AltAz transforms? I don't see it being in the former as not much has really changed there. 

> This is because the calls to erfa.apcs and calculation of the celestial-to-intermediate matrix are more expensive that the calls to erfa.apci we use in the master branch

This isn't true. Calculation of the celestial to intermediate matrix has always been required as it is needed by erfa.apci. Then there's the fact that erfa.apci ends up calling erfa.apcs anyway. 

If the slowdown is in the CIRS<->AltAz transformations, this is because of the call to `altaz_coo.location.get_gcrs_posvel()` necessitated by your choice of GCRS position and velocity to specify observer location for CIRS frames. While I completely understand the convenience and consistency of this approach, that call does require the costly calculation of the celestial to intermediate matrix. An alternate approach would be to use CIRS position and velocity for the observer location for CIRS frames. Then an `Earthlocation.get_cirs_posvel()` function could be written for use here that would be much less costly. Using CIRS position and velocity wouldn't be all that inconvenient either. In the GCRS<->CIRS transforms, it is necessary to calculate the celestial to intermediate matrix anyway, so this could be applied to the observer's position and velocity also during these transformations. In the ICRS<->CIRS transforms, the observer's GCRS position and velocity required by the call to erfa.apcs could also be easily obtained by using the celestial to intermediate matrix which is also calculated in `erfa_astrom.get().apci()` anyway.