@StuartLittlefair 2020-11-05
@mkbrewer, you are wrong about the cause of the slowdown, but then again so am I. 

The way it is coded is slightly slower but the bulk of the slowdown comes from the fact that CIRS->AltAz contains an ICRS self transform to get topocentric CIRS. 

So what was a fairly direct ICRS->CIRS->AA becomes ICRS->CIRS->ICRS->CIRS’->AA so there is double the number of actual transforms, each of which is slower.