@StuartLittlefair 2020-11-05
Looking into the slowdown further, there is something *very* odd going on here, which I'll have to dig into

```
import numpy as np
from astropy.time import Time
import astropy.units as u
from astropy.coordinates import EarthLocation, CIRS, AltAz, SkyCoord
from astropy.coordinates.tests.utils import randomly_sample_sphere

# lots of times
times = Time('2020-08-21T00:00') + np.linspace(-3*u.hour, 3*u.hour, 100)
location = EarthLocation.of_site('lapalma')

# aa frame at lots of times
aa_frame = AltAz(obstime=times[:, np.newaxis], location=location)
ra, dec, _ = randomly_sample_sphere(100)
coos = SkyCoord(ra, dec)

aa = coos.transform_to(aa_frame) # 3.2 sec (60 msec on master)
aa = coos.transform_to(CIRS(obstime=times)).transform_to(aa_frame) # 320 msec
```