@StuartLittlefair 2020-11-05
There is something about the speed here that I don't understand - if you make a large number of times and ICRS coordinates and broadcast them to make an altaz frame then the second line of code below is about 18x faster than the first... 

```python
coos.transform_to(aa_frame)
coos.transform_to(CIRS()).transform_to(aa_frame)
```