@mkbrewer 2020-11-05
> The way it is coded is slightly slower but the bulk of the slowdown comes from the fact that CIRS->AltAz contains an ICRS self transform to get topocentric CIRS.

Is this true in the case of a solar system body where one already has a topocentric GCRS position? Or is this just in the case where one is asking for the AltAz coordinates of an ICRS SkyCoord where the time and location are necessarily derived from the AltAz frame itself. In that case, couldn't you just derive the CIRS `obstime` and `location` from the AltAz frame before leaving the ICRS in the first place? You must already be doing that for`obstime` as an ICRS SkyCoord has no `obstime` attribute.