@mkbrewer 2020-11-05
Yeah, I seem to be consistently running one comment behind you. That (N, M) business is way out of my league, but it looks like you're on it.