@StuartLittlefair 2020-11-05
Well I know what the problem is, but not how to fix it! You are right about the cause of the slowdown for a single point though