@StuartLittlefair 2020-11-17
@maxnoe I think this is of similar performance now to before this PR. Running the test case you supplied in https://github.com/astropy/astropy/pull/10647#issuecomment-673569823 I see similar performance to astropy master:

master branch:

```
No Interpolation took 9.7546 s
Interpolation with  10.0 s   resolution took 1.0634 s
Interpolation with   5.0 min resolution took 0.2984 s
Interpolation with  30.0 min resolution took 0.2053 s
```

this pr:
```
No Interpolation took 10.8796 s
Interpolation with  10.0 s   resolution took 1.1008 s
Interpolation with   5.0 min resolution took 0.2722 s
Interpolation with  30.0 min resolution took 0.2027 s
```